renamed index files for navbar and footer
