+++ src/components/footerComponent/footer.js
@@ -23,3 +23,4 @@ class Navbar extends Component {
 }
 
 export default Navbar;
+
